arch: arm: config: add ADI rpi defconfigs
This adds our custom defconfig files for RPI. Before, we are always
creating a new patch to copy the configs from the previous active rpi
branch. Instead, let's just add it in master and the patch will be
automatically carried to the current rpi branch. When creating a new rpi
branch, this same patch will also be applied so we'll only have to
update the configs as we typically do.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>
(cherry picked from commit f695d3819825aac4376dcd546e97e81cf586f34c)