iio: adc: adrv9002: adi_adrv9001_cals: Error out if no enough coefficients
In case pour buffer is too short for the coefficients we need to set in
adi_adrv9001_cals_InitCals_WarmBoot_Coefficients_UniqueArray_Set(), we
return error. This makes sure that if someone loads a new
profile concurrently with loading the coefficients, we won't load a
subset of coefficients and proceed with the device initialization.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>