iio: adc: adrv9002: add warmboot support
Warmboot allows to bypass the Initial calibrations and instead write a
set of known-good calibration coefficients directly to memory. This
improves significantly boot and profile loading times.

The way this is done is by adding a new binary attribute called
'init_cals_coefficients'. The typical workflow looks like:

 1. cat init_cals_coefficients > current_profile_coeffs.bin
 2. cat current_profile_coeffs.bin > init_cals_coefficients
 3. cat current_profile.json > profile_config

So, looking at the above commands, one first needs to save the
coefficients for the desired profile (note that profile needs to be
loaded with warmboot disabled first so the initial calibrations can run
once). Then, life goes on and we might change to another profile but
when moving back to the same profile (now with warmboot enabled) one can
run 2 and 3 commands to load the saved coefficients and the profile.

If one just wants to move from a profile with warmboot disabled to the
exact same profile but with warmboot enabled, the following two commands
are enough::

 1. cat init_cals_coefficients > /dev/null
 2. cat current_profile.json > profile_config

The above is possible because the coefficients buffer is shared between
read/write (it's just too big to have more than one) which also has it's
downsides of course - concurrent calls on the read/write interfaces will
be a very bad idea (as the buffers are > PAGE_SIZE).

Signed-off-by: Nuno Sa <nuno.sa@analog.com>