spi: spi-axi-spi-engine: use max_speed_hz if no speed_hz given
If the speed is not given in the transfer, default to max. Mote this is
only needed because of the offload engine support where we bypass the
spi core. In the ideal worl, we should still call spi_validate() before
loading the offload message.

Something to keep in mind when upstreaming offload support.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>