ci: run-build.sh: don't build dts in exception file
If a devicetree is in the exception file, let's skip it. This should
speed up the job considerably and makes sure the job fails because some
other devicetrees (that we don't own) fail to build.

While on it, moved the 'hdl_project_tag_err' block so we can bail out
without building any devicetree if some hdl tag is missing. Note we were
failing the job anyways in this case.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>